# refactor(pointcloud_preprocessor): extract downsample logic from pickup_based_voxel_downsample_filter