--- a/sensing/autoware_pointcloud_preprocessor/include/autoware/pointcloud_preprocessor/downsample_filter/pickup_based_voxel_grid_downsample_filter_node.hpp
+++ b/sensing/autoware_pointcloud_preprocessor/include/autoware/pointcloud_preprocessor/downsample_filter/pickup_based_voxel_grid_downsample_filter_node.hpp
@@ -38,7 +38,7 @@ struct VoxelSize
 
 void downsample_with_voxel_grid(
   const sensor_msgs::msg::PointCloud2::ConstSharedPtr & input,
-  VoxelSize voxel_size,
+  const VoxelSize & voxel_size,
   sensor_msgs::msg::PointCloud2 & output
 );