## Description
- Extracted the downsampling logic into a new function, `autoware::pointcloud_preprocessor::downsample_with_voxel_grid()`.
- Refactored the point cloud casting logic using `PointCloud2ConstIterator`.
- Added corresponding unit tests for the new function.

## How was this PR tested?

### unit test performed
Added 5 new unit tests, all of which passed successfully.
```
8: [==========] Running 5 tests from 1 test suite.
8: [----------] Global test environment set-up.
8: [----------] 5 tests from DownsampleWithVoxelGridTest
8: [ RUN      ] DownsampleWithVoxelGridTest.TestBasicDownsampling
8: [       OK ] DownsampleWithVoxelGridTest.TestBasicDownsampling (0 ms)
8: [ RUN      ] DownsampleWithVoxelGridTest.TestEmptyPointCloud
8: [       OK ] DownsampleWithVoxelGridTest.TestEmptyPointCloud (0 ms)
8: [ RUN      ] DownsampleWithVoxelGridTest.TestSinglePoint
8: [       OK ] DownsampleWithVoxelGridTest.TestSinglePoint (0 ms)
8: [ RUN      ] DownsampleWithVoxelGridTest.TestDifferentVoxelSizes
8: [       OK ] DownsampleWithVoxelGridTest.TestDifferentVoxelSizes (0 ms)
8: [ RUN      ] DownsampleWithVoxelGridTest.TestNegativeCoordinates
8: [       OK ] DownsampleWithVoxelGridTest.TestNegativeCoordinates (0 ms)
8: [----------] 5 tests from DownsampleWithVoxelGridTest (0 ms total)
8: 
8: [----------] Global test environment tear-down
8: [==========] 5 tests from 1 test suite ran. (0 ms total)
8: [  PASSED  ] 5 tests.
```

### performance test

To verify performance, I measured the CPU usage of the `pickup_based_downsample_filter_node` using `perf` while replaying a LiDAR PointCloud2 rosbag.

The rosbag used for testing can be found here:
https://drive.google.com/drive/folders/1Hh25JXA7qi3tlOhY5NNs8YNsxZjrp02l?usp=drive_link
This rosbag was generated from a topic in the logging simulator tutorial:
https://autowarefoundation.github.io/autoware-documentation/main/tutorials/ad-hoc-simulation/rosbag-replay-simulation/

The test was run using the following commands:
```
# termianl1
cd ~/autoware && colcon build --symlink-install --packages-select autoware_pointcloud_preprocessor && source install/local_setup.bash  && ros2 run autoware_pointcloud_preprocessor pickup_based_voxel_grid_downsample_filter_node --ros-args -p voxel_size_x:=0.1 -p voxel_size_y:=0.1 -p voxel_size_z:=0.1 -r input:=/sensing/lidar/top/pointcloud

# termianl2
source ~/autoware/install/local_setup.bash && ros2 bag play ~/autoware_map/sample-pointcloud-rosbag/

# terminal3
ros2 topic hz /output -w 10

# terminal3
sudo perf stat -p $(pgrep -d, -f pickup_based_voxel_grid_downsample_filter_node) sleep 10
```

Here are the CPU usage results from three separate runs:

| Condition | 1st | 2nd | 3rd |
| :--- | :--- | :--- | :--- |
| Before | 0.228 | 0.231 | 0.224 |
| After | 0.223 | 0.224 | 0.228 |

No performance degradation was observed.

## Notes for reviewers

The previous integration tests from PR [#11003](https://github.com/autowarefoundation/autoware_universe/pull/11003) were removed. 
They were time-consuming and largely overlapped with the scope of the new unit tests introduced in this PR.


## Interface changes

None.

## Effects on system behavior

None.